BE-Feat/#27-rest template

See merge request s07-bigdata-recom-sub2/S07P22A503!41

+++ server/src/main/java/com/singsong/db/entity/Tag.java
@@ -17,7 +17,7 @@ import javax.persistence.*;
 public class Tag {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int tagId;
+    Long tagId;
 
     @Column(length = 20)
     String tagName;
